pyproject.toml
[tool.poetry]
name = "lims-utils"
version = "0.1.0"
description = ""
authors = ["Brandon Davis <git@subdavis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
youtube-transcript-api = "^0.6.2"
python-dateutil = "^2.9.0.post0"
types-python-dateutil = "^2.9.0.20240821"
anthropic = "^0.34.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
types-requests = "^2.32.0.20240914"
python-youtube = "^0.9.6"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
mypy = "^1.11.2"
isort = "^5.13.2"
black = "^24.8.0"

[tool.poetry.scripts]
updateCache = "lims_utils.scripts:updateCache"
generate = "lims_utils.scripts:generate"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true

[tool.isort]
profile = "black"